refactor: move resource methods to separate model/interactor
