+++ src/me/dc37/smp/controllers/SplashScreenController.java
@@ -2,25 +2,25 @@ package me.dc37.smp.controllers;
 
 import javafx.scene.image.Image;
 import javafx.scene.layout.Region;
-import me.dc37.smp.interactors.SMPAppInteractor;
+import me.dc37.smp.interactors.ResourceInteractor;
 import me.dc37.smp.interactors.SplashScreenInteractor;
-import me.dc37.smp.models.SMPAppModel;
+import me.dc37.smp.models.ResourceModel;
 import me.dc37.smp.models.SplashScreenModel;
 import me.dc37.smp.views.SplashScreenViewBuilder;
 
 public class SplashScreenController {
 
-	private final SMPAppModel mwModel;
-	private final SMPAppInteractor mwInteractor;
-	
+    private final ResourceModel resModel;
+    private final ResourceInteractor resInteractor;
+    
 	private final SplashScreenModel model;
 	private final SplashScreenInteractor interactor;
 	private final SplashScreenViewBuilder viewBuilder;
 	
 	public SplashScreenController() {
-		mwModel = SMPAppModel.getInstance();
-		mwInteractor = new SMPAppInteractor(mwModel);
-		
+	    resModel = ResourceModel.getInstance();
+	    resInteractor = new ResourceInteractor(resModel);
+	    
 		model = new SplashScreenModel();
 		interactor = new SplashScreenInteractor(model);
 		
@@ -32,8 +32,8 @@ public class SplashScreenController {
 	}
 	
 	public Image getHeaderIcon() {
-		mwInteractor.setHeaderIcon();
-		return mwModel.getHeaderIcon();
+		resInteractor.setHeaderIcon();
+		return resModel.getHeaderIcon();
 	}
 	
 	public String getTitle() {
